requirements.txt
Django==5.1.1
djangorestframework==3.15.2
django-filter==25.1
drf-spectacular==0.28.0
psycopg2-binary==2.9.9
Pillow==11.1.0
# environs==14.2.0
# requests==2.32.3
python-dotenv==1.0.1
